feat: Рефакторинг кода на React

+++ src/utils/Api.js
@@ -64,22 +64,22 @@ _checkResponse(res) {
       .then(this._checkResponse)
   }
 
-  deleteLike(id) {
-    return fetch(
-      `${this._baseUrl}/cards/${id}/likes`, {
-        method: "DELETE",
-        headers: this._headers,
-      })
-      .then(this._checkResponse)
-  }
-
-  addLike(id) {
-    return fetch(
-      `${this._baseUrl}/cards/${id}/likes`, {
-        method: "PUT",
-        headers: this._headers,
-      })
-      .then(this._checkResponse)
+  changeLikeCardStatus(id, checkLikeVariable) {
+    if (checkLikeVariable) {
+      return fetch(
+        `${this._baseUrl}/cards/${id}/likes`, {
+          method: "PUT",
+          headers: this._headers,
+        })
+        .then(this._checkResponse)
+    } else {
+      return fetch(
+        `${this._baseUrl}/cards/${id}/likes`, {
+          method: "DELETE",
+          headers: this._headers,
+        })
+        .then(this._checkResponse)
+    }
   }
 
   changeProfileAvatar(avatar) {
